Refactor ‘client.utils’ into a package
