+++ linnote/client/utils/__init__.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+u"""
+Miscellaneaous tools for the client.
+
+Author: Anatole Hanniet, Tutorat Santé Lyon Sud (2014-2017).
+License: Mozilla Public License, see 'LICENSE.txt' for details.
+"""
+
+from .login import LOGIN_MANAGER
+from .session import session
